refactor: npc_commands.py 3개 클래스 → npc/ 디렉토리 분리 (태스크 29.4)
- npc/talk_command.py, npc/trade_command.py, npc/shop_command.py
- npc/__init__.py re-export, npc_commands.py 하위 호환성 유지
