+++ src/mud_engine/commands/npc/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""NPC 명령어 패키지"""
+from .talk_command import TalkCommand
+from .trade_command import TradeCommand
+from .shop_command import ShopCommand
+
+__all__ = ['TalkCommand', 'TradeCommand', 'ShopCommand']
